Refactor utils into separate file
